move filter types to types.d.ts
